Major discovery in forums for creating a source
Just creating the source and the settings to apply to it is not enough
for obs to recognize the source as one of the known types. To make it so
you have to give obs_source_create an ID that comes from a list of known
literals, one of which is "game_capture".