Various changes
* Souces are not as easy to handle as scenes, so instead of
find_or_create* function, I opted for an exists function and if False,
we go ahead and create the scene. Plain and simple.
* Removed all global state. It's better to just reference and release
references in any function since it avoid problems with which functions
change ref count and which don't and where we need to release them.
* Opted for two functions to set most things up on frontend finished
loading event. One function creates the scene and the source. The other
sets up signal handlers to handle events like obs finding the window of
a newly opened game process, or losing the connection when it closes.