Merged all py files into ags_qa
Python and obs-studio are annoying in that you can't develop in multiple
files and still reference only the one you need. There's also no way
to safely merge them in a on-save step in a different folder which is
what obs-studio uses, so we always update a single file but obs still
gets the whole thing updated. So to fix that, I moved all code into a
single file and moved all static functions under classes with the
filename as the namespace and marked them all static. This way we can
collapse them all easily and avoid scrolling forever.