Improvements on signal handling
* The two other scripts by Penwywern are referenced for this change.
We basically have two places we need to check for signal connections
On script_load since we load scripts before the sources (this covers the
case where the source is already created from previous sessions) and
also on new source creation(e.g. the first time we create the scene and
source  we'll use). That's when we need to get the signal handler of the
source and connect to hook/unhook events. As for the eye icon, we need
the scene_item's item_visible signal. It's basically the scene's. So we
can connect to them on the scene's (as a source) signal handler. For the
source_create/destroy/rename signals, we can use the global signal
handler. We disconnect either on script_unload or source_destroy etc.