Fixed two things in order to revert the bad state unhook's callback created
* Removed the while(True). This was added because the crash check is
done by checking the state of the two windows of the game, the main and
the crash. But there's a time window of a couple seconds when it
transitions from running to crashed state where both window handles are
invalid. But the way it was written, it did not account for cases where
the app is not supposed to end during unhook.
* Added checks for that code to run only when the process is still
running and also set proc var to None only if it crashed or exited.